# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Alba Digital API",
        "description": "An API of ecommerce",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Category"
                ],
                "summary": "Paginated category list",
                "description": "Paginated list of categories",
                "operationId": "categoryList",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "number",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "category 1"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "example": "2023-08-28 06:06:17"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/user/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "User login",
                "operationId": "login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "john@gmail.com",
                                    "password": "helloworld"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Valid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@test.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "api_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNTVmNjA1NDNmMGI1MDg3MGFlNWY5ZWQ1ZTMzYWIwNzhiZTQ1NjRmYzliMmMyMjQwZjNjNDFlZWQ4ODhiZDcxMTg3NWJiM2E0NWNhY2IxMmMiLCJpYXQiOjE2OTI1NDg3MzAuODU3OTE4LCJuYmYiOjE2OTI1NDg3MzAuODU3OTIzLCJleHAiOjE3MjQxNzExMzAuNzczNTQ5LCJzdWIiOiIxIiwic2NvcGVzIjpbXX0.JBAMuBjIqP78z_6BzubosjNfk0O5TCR1Ro73y6wSJPODLbMyxsQWFJmqroSHVYzIHpfC9ikLIUz4ywKCrsuqyOgWzhnaQH8W-A5eqnfbp-Vrevq9QzS2cGVTGySYa_-YxP3sy8fgEfMApDKzmSDJZCy-Q61aYRYjtp3v52Nj6iAd_mvrFQ8TeVq-Vbor2yTXGZnnd7SHTJxzR4eDCmg3RvhI-4vKDGF4LGRt8Q5bQOK0tIY8bvsEpsXDVlkNlSYEu0txOFYkYjNqaxnBDV9BTF3Wki6Ui6qc5BP9EwvtGugerygdT0LDQoblWHXx9gxKBB4-Y-CN4xYa4rvTIeGWEioqPHpywINtWmNWVJ71gheva8aBIc0sxNU3i2xa6dT5JF_rEorHXsUPkyXjGQzxMUfTP9pmtKlbq1eW-AJuS23S_7p1g-CNYMvx2hMtNeDSKYCQE-RZoayCEiLJS10ldWZcF-iMgEm-ZcTG8lzjjfYdqRAMS7FyAM6f1xmAvhgnSDPW-nZnbGo5AUueT3Zz3WrvU6w4BdtiwVLdz15bEmJ3ZkT2x6GvJIqx2DQNhBO9x6o6kVrujcdtc_JDF9tkPNsu171tiV6Uk8kV_r15jyO7taSugT4uqS1ITdd4II4oAKL0JXRd4pSVsY8Fbr0p-HOdYmebLHsnUGRmMiDiZBA"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Incorrect username or password!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register",
                "operationId": "0a38310f833466b1b4d00ef8d0a449b9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "email": {
                                                "type": "string"
                                            },
                                            "password": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "John",
                                    "email": "john@gmail.com",
                                    "password": "helloworld"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "number",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John"
                                                },
                                                "email": {
                                                    "type": "string",
                                                    "example": "john@test.com"
                                                },
                                                "email_verified_at": {
                                                    "type": "string",
                                                    "example": null
                                                },
                                                "api_token": {
                                                    "type": "string",
                                                    "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiJ9.eyJhdWQiOiIxIiwianRpIjoiNTVmNjA1NDNmMGI1MDg3MGFlNWY5ZWQ1ZTMzYWIwNzhiZTQ1NjRmYzliMmMyMjQwZjNjNDFlZWQ4ODhiZDcxMTg3NWJiM2E0NWNhY2IxMmMiLCJpYXQiOjE2OTI1NDg3MzAuODU3OTE4LCJuYmYiOjE2OTI1NDg3MzAuODU3OTIzLCJleHAiOjE3MjQxNzExMzAuNzczNTQ5LCJzdWIiOiIxIiwic2NvcGVzIjpbXX0.JBAMuBjIqP78z_6BzubosjNfk0O5TCR1Ro73y6wSJPODLbMyxsQWFJmqroSHVYzIHpfC9ikLIUz4ywKCrsuqyOgWzhnaQH8W-A5eqnfbp-Vrevq9QzS2cGVTGySYa_-YxP3sy8fgEfMApDKzmSDJZCy-Q61aYRYjtp3v52Nj6iAd_mvrFQ8TeVq-Vbor2yTXGZnnd7SHTJxzR4eDCmg3RvhI-4vKDGF4LGRt8Q5bQOK0tIY8bvsEpsXDVlkNlSYEu0txOFYkYjNqaxnBDV9BTF3Wki6Ui6qc5BP9EwvtGugerygdT0LDQoblWHXx9gxKBB4-Y-CN4xYa4rvTIeGWEioqPHpywINtWmNWVJ71gheva8aBIc0sxNU3i2xa6dT5JF_rEorHXsUPkyXjGQzxMUfTP9pmtKlbq1eW-AJuS23S_7p1g-CNYMvx2hMtNeDSKYCQE-RZoayCEiLJS10ldWZcF-iMgEm-ZcTG8lzjjfYdqRAMS7FyAM6f1xmAvhgnSDPW-nZnbGo5AUueT3Zz3WrvU6w4BdtiwVLdz15bEmJ3ZkT2x6GvJIqx2DQNhBO9x6o6kVrujcdtc_JDF9tkPNsu171tiV6Uk8kV_r15jyO7taSugT4uqS1ITdd4II4oAKL0JXRd4pSVsY8Fbr0p-HOdYmebLHsnUGRmMiDiZBA"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2023-08-28 06:06:17"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Username and Password did not match"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/user/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "operationId": "e549fd8544c96edfed36dd13c027de7e",
                "responses": {
                    "201": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}